move `{scaled,raw}_to_{raw,scaled}()` to SignalBase
