+++ cantools/database/can/signal.py
@@ -1,10 +1,8 @@
 # A CAN signal.
-import contextlib
 import decimal
 from typing import TYPE_CHECKING, List, Optional, Union
 
 from ...typechecking import ByteOrder, Choices, Comments, SignalValueType
-from ..namedsignalvalue import NamedSignalValue
 from ..signalbase import SignalBase
 
 if TYPE_CHECKING:
@@ -225,49 +223,6 @@ class Signal(SignalBase):
             # assume that we have either no comment at all or a
             # multilingual dictionary
             self.comments = comment
-
-    def raw_to_scaled(
-        self, raw: Union[int, float], decode_choices: bool = True
-    ) -> SignalValueType:
-        """Convert an internal raw value according to the defined scaling or value table.
-
-        :param raw:
-            The raw value
-        :param decode_choices:
-            If `decode_choices` is ``False`` scaled values are not
-            converted to choice strings (if available).
-        :return:
-            The calculated scaled value
-        """
-        if decode_choices:
-            with contextlib.suppress(KeyError, TypeError):
-                return self.choices[raw]  # type: ignore[index]
-
-        if self.offset == 0 and self.scale == 1:
-            # treat special case to avoid introduction of unnecessary rounding error
-            return raw
-        return raw * self.scale + self.offset
-
-    def scaled_to_raw(self, scaled: SignalValueType) -> Union[int, float]:
-        """Convert a scaled value to the internal raw value.
-
-        :param scaled:
-            The scaled value.
-        :return:
-            The internal raw value.
-        """
-        if isinstance(scaled, (float, int)):
-            _transform = float if self.is_float else round
-            if self.offset == 0 and self.scale == 1:
-                # treat special case to avoid introduction of unnecessary rounding error
-                return _transform(scaled)  # type: ignore[operator,no-any-return]
-
-            return _transform((scaled - self.offset) / self.scale)  # type: ignore[operator,no-any-return]
-
-        if isinstance(scaled, (str, NamedSignalValue)):
-            return self.choice_string_to_number(str(scaled))
-
-        raise TypeError(f"Conversion of type {type(scaled)} is not supported.")
 
     @property
     def comment(self) -> Optional[str]:
