+++ cantools/database/signalbase.py
@@ -1,10 +1,11 @@
+import contextlib
 import logging
 from typing import List, Optional, Tuple, Union
 
-from ..typechecking import ByteOrder, Choices, Interval
+from ..typechecking import ByteOrder, Choices, Interval, SignalValueType
+from .namedsignalvalue import NamedSignalValue
 
 logger = logging.getLogger(__name__)
-
 
 class SignalBase:
     def __init__(
@@ -198,3 +199,46 @@ class SignalBase:
     def scale(self, new_scale: Union[float, List[float]]) -> None:
         """Set scale"""
         self._scale = new_scale
+
+    def raw_to_scaled(
+        self, raw: Union[int, float], decode_choices: bool = True
+    ) -> SignalValueType:
+        """Convert an internal raw value according to the defined scaling or value table.
+
+        :param raw:
+            The raw value
+        :param decode_choices:
+            If `decode_choices` is ``False`` scaled values are not
+            converted to choice strings (if available).
+        :return:
+            The calculated scaled value
+        """
+        if decode_choices:
+            with contextlib.suppress(KeyError, TypeError):
+                return self.choices[raw]  # type: ignore[index]
+
+        if self.offset == 0 and self.scale == 1:
+            # treat special case to avoid introduction of unnecessary rounding error
+            return raw
+        return raw * self.scale + self.offset
+
+    def scaled_to_raw(self, scaled: SignalValueType) -> Union[int, float]:
+        """Convert a scaled value to the internal raw value.
+
+        :param scaled:
+            The scaled value.
+        :return:
+            The internal raw value.
+        """
+        if isinstance(scaled, (float, int)):
+            _transform = float if self.is_float else round
+            if self.offset == 0 and self.scale == 1:
+                # treat special case to avoid introduction of unnecessary rounding error
+                return _transform(scaled)  # type: ignore[operator,no-any-return]
+
+            return _transform((scaled - self.offset) / self.scale)  # type: ignore[operator,no-any-return]
+
+        if isinstance(scaled, (str, NamedSignalValue)):
+            return self.choice_string_to_number(str(scaled))
+
+        raise TypeError(f"Conversion of type {type(scaled)} is not supported.")
